k8s-windows-linux/code/requirements.txt
azure-storage
argparse
